backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic==2.11.7
scikit-learn==1.7.1
pandas==2.3.1
joblib==1.5.1
python-multipart==0.0.20
